Check if the host supports 64-bit processes before attempting to start the x64 helper process.
This change adds a function, canRun64BitPrograms() to OverlayWinPrivate.

If the program is built for x64, the function unconditionally
returns true.

If the program is built for x86, the function uses a call to the
runtime-resolved kernel32!IsWow64Process to determine whether or
not the system is x64 capable.

This functionality is needed to avoid spawning the x64 helper in
case the system cannot support it. Doing so without care, like we
did previously, could be fatal because Mumble immediately restarts
the helper process if it terminates. This would presumably make the
host computer unusable.

+++ src/mumble/Overlay_win.cpp
@@ -50,7 +50,30 @@
 // mumble_app.dll world.
 extern "C" __declspec(dllexport) void mumbleSelfDetection() {};
 
+// Determine if the current Mumble client is able to host
+// x64 programs.
+//
+// If we're on x86, we use use the IsWoW64Process function
+// to determine this.  If we're on x64, we already know we're
+// capable, so we simply return TRUE.
+static bool canRun64BitPrograms() {
+#if defined(_M_X64)
+	return TRUE;
+#elif defined(_M_IX86)
+	typedef BOOL (WINAPI *IsWow64ProcessPtr)(HANDLE, BOOL *);
+	IsWow64ProcessPtr wow64check = (IsWow64ProcessPtr)GetProcAddress(GetModuleHandle(L"kernel32"), "IsWow64Process");
+	if (wow64check) {
+		BOOL isWoW64 = FALSE;
+		wow64check(GetCurrentProcess(), &isWoW64);
+		return isWoW64;
+	}
+	return FALSE;
+#endif
+}
+
 OverlayPrivateWin::OverlayPrivateWin(QObject *p) : OverlayPrivate(p) {
+	m_allow64bit = canRun64BitPrograms();
+
 	m_helper_exe_path = QString::fromLatin1("%1/mumble_ol.exe").arg(qApp->applicationDirPath());
 	m_helper_exe_args = QStringList(QString::number(OVERLAY_MAGIC_NUMBER));
 	m_helper_process = new QProcess(this);
@@ -94,10 +117,14 @@ void OverlayPrivateWin::setActive(bool active) {
 
 		if (m_active) {
 			startHelper(m_helper_process);
-			startHelper(m_helper64_process);
+			if (m_allow64bit) {
+				startHelper(m_helper64_process);
+			}
 		} else {
 			m_helper_process->terminate();
-			m_helper64_process->terminate();
+			if (m_allow64bit) {
+				m_helper64_process->terminate();
+			}
 		}
 	}
 }
